WIP Impl of Compresser Builder
- DEV: Impl of compresser builder layer
- DEV: Created class to hold compresser configuration

+++ src/main/java/br/com/xavier/zipper/interfaces/compression/IZipperCompresserBuilder.java
@@ -7,7 +7,7 @@ import java.util.zip.Checksum;
 public interface IZipperCompresserBuilder extends Serializable {
 	
 	IZipperCompresserBuilder buffer(OutputStream os);
-	IZipperCompresserBuilder bytesPerRead( Long bytesPerRead );
+	IZipperCompresserBuilder bytesPerRead( Integer bytesPerRead );
 	IZipperCompresserBuilder checksum(Checksum checksum);
 	
 	IZipperCompresser build();
